move methodology and translations to TenantStore
