+++ src/RootStore/__tests__/TenantStore.test.js
@@ -19,7 +19,7 @@ import createAuth0Client from "@auth0/auth0-spa-js";
 
 import RootStore from "../RootStore";
 import { CURRENT_TENANT_IN_SESSION } from "../TenantStore/TenantStore";
-import { US_MO } from "../../views/tenants/utils/lanternTenants";
+import { US_MO } from "../TenantStore/lanternTenants";
 import { doesUserHaveAccess, getAvailableStateCodes } from "../utils/user";
 import { METADATA_NAMESPACE } from "../../constants";
