+++ src/components/__tests__/Revocations.test.js
@@ -31,7 +31,7 @@ import CaseTable from "../charts/new_revocations/CaseTable/CaseTable";
 import { METADATA_NAMESPACE } from "../../constants";
 import { setTranslateLocale } from "../../views/tenants/utils/i18nSettings";
 
-import { US_MO } from "../../views/tenants/utils/lanternTenants";
+import { US_MO } from "../../RootStore/TenantStore/lanternTenants";
 import mockWithTestId from "../../../__helpers__/mockWithTestId";
 import filterOptionsMap from "../../views/tenants/constants/filterOptions";
 import {
