+++ src/components/charts/new_revocations/CaseTable/utils/__tests__/helpers.test.js
@@ -16,7 +16,7 @@
 // =============================================================================
 import * as helpers from "../helpers";
 import { setTranslateLocale } from "../../../../../../views/tenants/utils/i18nSettings";
-import * as lanternTenant from "../../../../../../views/tenants/utils/lanternTenants";
+import * as lanternTenant from "../../../../../../RootStore/TenantStore/lanternTenants";
 
 describe("helper", () => {
   let data;
