+++ src/utils/downloads/downloadCanvasAsImage.js
@@ -48,6 +48,7 @@ function downloadCanvasAsImage({
   chartId,
   timeWindowDescription,
   shouldZipDownload,
+  methodology,
 }) {
   const imageData = transformCanvasToBase64(canvas, chartTitle, filters);
 
@@ -56,7 +57,8 @@ function downloadCanvasAsImage({
       chartId,
       chartTitle,
       timeWindowDescription,
-      filters
+      filters,
+      methodology
     );
 
     const imageFile = {
